remove draft terminology
