+++ deployables/app/app/routes/roles/list-layout.tsx
@@ -28,12 +28,6 @@ const RoleListLayout = ({ params: { workspaceId } }: Route.ComponentProps) => {
           </TabBar.LinkTab>
 
           <TabBar.LinkTab
-            to={href('/workspace/:workspaceId/roles/drafts', { workspaceId })}
-          >
-            Drafts
-          </TabBar.LinkTab>
-
-          <TabBar.LinkTab
             to={href('/workspace/:workspaceId/roles/on-chain', { workspaceId })}
           >
             On-Chain only
